change time_t to int64_t to support long values for seconds and milliseconts

+++ libc/include/dtime.h
@@ -28,8 +28,8 @@ extern "C" {
 
 #include <stdlib.h>
 
-typedef uint32_t time_t;
-typedef uint32_t suseconds_t;
+typedef int64_t time_t;
+typedef int64_t suseconds_t;
 typedef int clock_t;
 
 typedef struct {
